+++ iOS/AreUDone/AreUDone/Service/ListService/ListService.swift
@@ -11,12 +11,12 @@ import NetworkFramework
 protocol ListServiceProtocol {
   func createList(withBoardId boardId: Int, title: String, completionHandler: @escaping (Result<Void, APIError>) -> Void)
   func deleteList(withListId listId: Int, completionHandler: @escaping (Result<Void, APIError>) -> Void)
-  func updateList(withListId listId: Int, position: Int?, title: String?, completionHandler: @escaping (Result<Void, APIError>) -> Void)
+  func updateList(withListId listId: Int, position: Double?, title: String?, completionHandler: @escaping (Result<Void, APIError>) -> Void)
 }
 
 extension ListServiceProtocol {
 
-  func updateList(withListId listId: Int, position: Int? = nil, title: String? = nil, completionHandler: @escaping (Result<Void, APIError>) -> Void) {
+  func updateList(withListId listId: Int, position: Double? = nil, title: String? = nil, completionHandler: @escaping (Result<Void, APIError>) -> Void) {
     updateList(withListId: listId, position: position, title: title, completionHandler: completionHandler)
   }
 }
@@ -49,7 +49,7 @@ class ListService: ListServiceProtocol {
     }
   }
   
-  func updateList(withListId listId: Int, position: Int?, title: String?, completionHandler: @escaping (Result<Void, APIError>) -> Void) {
+  func updateList(withListId listId: Int, position: Double?, title: String?, completionHandler: @escaping (Result<Void, APIError>) -> Void) {
     router.request(route: ListEndPoint.updateList(listId: listId, position: position, title: title)) { result in
       completionHandler(result)
     }
